Chore: 마이페이지 라우트 구조 수정
